refactor: extract SQL queries into a separate queries.go file for better organization and maintainability
